requirements.txt
# requirements.txt

numpy==1.26.4
pillow==10.3.0
pybind11==2.12.0
pyglet==2.0.15
